Wire up data endpoint
We'll access this `/data` endpoint to view aggregated salary
submissions. Our goal is to fill out this view with a line chart of
salary over years of experience, and other details including: current
average, minimum, and maximum salaries by graduation year.

This paves the way for this PRs goal to add the line chart.

Issue: 18