Add salary line chart to /data page
Adds a downloadable Charts.js line chart to display average, min, and
max salaries on our /data page.

Includes a chart service object that will be responsible for
formatting our data so that Chartkick can craft it into charts.

Starts this chart service with a method to format aggregate data for
a salaries by years of experience chart.

Issue: 18