Refactor line chart component parameters
Adds parameters to our line chart component so that we can we reuse
it more readily. This supports defaults as well as additional options
to be passed in, since Chartkick supports more options and each graph
may have different specifics.

This sets up specialized_options is an optional parameter, since
its contents are extra specifics that are not required to craft a
chart. This includes some preliminary defaults that we can update as
needed.

Note that Chartkick receives options as variable keyword arguments,
which is why we're passing them in with a double splat operator (**)
to Chartkick's #line_chart method.

Link: https://github.com/ankane/chartkick/blob/master/lib/chartkick/helper.rb
Issue: 18