Salary line chart (#83)
* Wire up data endpoint

We'll access this `/data` endpoint to view aggregated salary
submissions. Our goal is to fill out this view with a line chart of
salary over years of experience, and other details including: current
average, minimum, and maximum salaries by graduation year.

This paves the way for this PRs goal to add the line chart.

Issue: 18

* Add Chartkick

This gem will help us craft beautiful JavaScript charts for visualizing
salary data. It currently has lots of users, good support & docs, and
recent updates.

Chartkick is a wrapper around several charting libraries, including
Charts.js, Google Charts, and HighCharts. This initializes Chartkick to
pair with Charts.js.

Link: https://github.com/ankane/chartkick
Issue: 18

* Add salary line chart to /data page

Adds a downloadable Charts.js line chart to display average, min, and
max salaries on our /data page.

Includes a chart service object that will be responsible for
formatting our data so that Chartkick can craft it into charts.

Starts this chart service with a method to format aggregate data for
a salaries by years of experience chart.

Issue: 18

* Refactor line chart component parameters

Adds parameters to our line chart component so that we can we reuse
it more readily. This supports defaults as well as additional options
to be passed in, since Chartkick supports more options and each graph
may have different specifics.

This sets up specialized_options is an optional parameter, since
its contents are extra specifics that are not required to craft a
chart. This includes some preliminary defaults that we can update as
needed.

Note that Chartkick receives options as variable keyword arguments,
which is why we're passing them in with a double splat operator (**)
to Chartkick's #line_chart method.

Link: https://github.com/ankane/chartkick/blob/master/lib/chartkick/helper.rb
Issue: 18

* Run standardrb linter to make circle happy

Issue: 18